% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{NetBIDshiny.run4MR}
\alias{NetBIDshiny.run4MR}
\title{\code{NetBIDshiny.run4MR} is a function to run a shiny app for NetBID2 master regulator identification.}
\usage{
NetBIDshiny.run4MR()
}
\description{
\code{NetBIDshiny.run4MR} is a shiny app for NetBID2 master regulator 
identification and the generation of figures for top drivers with the help of two lazy functions in NetBID2,
NetBID.lazyMode.DriverEstimation() and NetBID.lazyMode.DriverVisualization(). 
No option is required. 
User could follow the online tutorial \url{https://jyyulab.github.io/NetBID_shiny/} for usage.
}
