% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{NetBIDshiny.run4Vis}
\alias{NetBIDshiny.run4Vis}
\title{\code{NetBIDshiny.run4Vis} is a function to run a shiny app for NetBID2 result visualization.}
\usage{
NetBIDshiny.run4Vis()
}
\description{
\code{NetBIDshiny.run4Vis} is a shiny app for NetBID2 result visualization. 
No option is required. 
User could follow the online tutorial \url{https://jyyulab.github.io/NetBID_shiny/} for usage.
}
