% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_NetBID_shiny}
\alias{run_NetBID_shiny}
\title{\code{run_NetBID_shiny} is a function to run a shiny app for NetBID2 result visualization.}
\usage{
run_NetBID_shiny()
}
\description{
\code{run_NetBID_shiny} is a shiny app for NetBID2 result visualization. No option is required. 
User could follow the online tutorial \url{https://jyyulab.github.io/NetBID_shiny/} for usage.
}
